Add shared product generator
